Added in Unit Tests and minor upgrades and refactoring

+++ mcCodingChallenge/mcCodingChallenge/Model/Crypto.swift
@@ -10,7 +10,7 @@ import Foundation
 class Crypto: Decodable {
     let name: String
     let symbol: String
-    let value: Double
+    let value: Decimal
     let imageUrl: String
     
     enum CodingKeys: String, CodingKey {
@@ -20,7 +20,7 @@ class Crypto: Decodable {
         case imageUrl = "image"
     }
     
-    init(name: String, symbol: String, value: Double, imageUrl: String) {
+    init(name: String, symbol: String, value: Decimal, imageUrl: String) {
         self.name = name
         self.symbol = symbol
         self.value = value
